requirements.txt
requests
beautifulsoup4
anthropic
schedule
python-dateutil
python-dotenv
openai
caldav
icalendar
